backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.8
inflection==0.5.1
mailgun==0.1.1
mailjet-rest==1.3.4
multidict==6.0.5
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pyactiveresource==1.0.1
PyJWT==2.9.0
python-http-client==3.3.7
pytz==2024.1
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
sendgrid==6.11.0
setuptools==73.0.1
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
twilio==9.2.4
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4